Introduce basis struct
It makes change of convention easier when for instance spherical
coordinates formula works with xpos as azimuth reference and zpos as
zenith but you want to use it with zneg as azimuth reference and xneg as
zenith. Convertion code between bases is almost completely optimized out
at compile time.
Also :
 - remove section common 3D vectors from constants since basis uses
 struct xpos, ypos, etc... and make the others in constant useless.
 - fix typo c_half_p -> c_half_pi